==> qmd/03_real-dends.qmd <==
---
    title: "Discovering Focused Microgenre Communities: Dendrograms"
    author: "Omar Lizardo"
    date: "`r Sys.Date()`"
---

# Setup
```{r setup}
    knitr::opts_chunk$set(include=FALSE, echo = TRUE, warning = FALSE, message = FALSE)
    library(dendextend)
    library(here)
    library(RColorBrewer)
    library(tidyverse)
```

```{r Loading link distance matrix and performing link clustering}
    load(file= here("dat", "lclust-res.RData"))
```

```{r Performing Link Clustering}
    lclust.res <- hclust(ldist.res$dist, method = "ward.D2") #clustering of link by link distance matrix
    lclust.dend <- as.dendrogram(lclust.res) #transforming link clustering into dendrogram
    dend.macro <- cut(lclust.dend, h = 8) #cutting at macrogenre level
    a <- data.frame(g.size = sapply(dend.macro$lower, nleaves)) #macrogenres as ordered in dendrogram results
    b <- data.frame(g.size = colSums(taste.dat[, 2:21]), 
                    g.names = names(colSums(taste.dat[, 2:21]))) #original macrogenre sizes
    c <- right_join(a, b) #joining data frames by macrogenre sizes to get macrogenre labels ordered by dendrogram results
    names(dend.macro$lower) <- c$g.names #naming macrogenre dendrogram list
```

# Dendrogram Plots from Link Clustered Data
```{r Creating macrogenre dendrograms}
    d1 <- lclust.dend %>% 
      set("labels", c(rep("", 20))) %>% 
      set("branches_k_color", value = brewer.pal(20, "Dark2"), k = 20) %>% 
      set("branches_lwd", value = 3) 

    d2 <- dend.macro$upper %>% 
      set("labels", dend.macro$lower %>% names) %>% 
      set("branches_k_color", value = brewer.pal(20, "Dark2")) %>% 
      set("branches_lwd", value = 4) %>% 
      set("labels_color", value = brewer.pal(20, "Dark2")) %>% 
      set("labels_cex", value = 2.5)

```

```{r Plotting macrogenre dendrograms}
    dev.off()
    png(here("Plots", "Dend", "macro.png"), height = 900, width = 1600)
    plot(d1) %>% 
      abline(h = 8, col="red", lty = 2, lwd = 3)
    
    dev.off()
    png(here("Plots", "Dend", "macro-lab.png"), height = 900, width = 1600)
    plot(d2) 
```

```{r Creating microgenre dendrograms}
    d3 <- dend.macro$lower$Metal %>% 
      set("labels", c(rep("", nleaves(dend.macro$lower$Metal)))) %>% 
      set("branches_lwd", value = 4) %>% 
      set("branches_k_color", value = brewer.pal(3, "Dark2"), k = 3)
    
    d4 <- dend.macro$lower$Salsa %>% 
      set("labels", c(rep("", nleaves(dend.macro$lower$Salsa)))) %>% 
      set("branches_lwd", value = 4) %>% 
      set("branches_k_color", value = brewer.pal(3, "Dark2"), k = 3)
```


```{r Plotting microgenre dendrograms}
    dev.off()
    png(here("Plots", "Dend", "metal-branches.png"), height = 900, width = 1600)
    plot(d3, axes = FALSE) 
      par(cex = 5) 
      title(xlab = "", ylab = "", main = "Metal")
    
    dev.off()
    png(here("Plots", "Dend", "salsa-branches.png"), height = 900, width = 1600)
    plot(d4, axes = FALSE)  
      par(cex = 5) 
      title(xlab = "", ylab = "", main = "Salsa")
```
