==> qmd/04_microgenre-analysis.qmd <==
---
  title: "Discovering Focused Microgenre Communities: Microgenre Analysis"
  author: "Omar Lizardo"
  date: "`r Sys.Date()`"
  output:
    html_document: default
---

# Setup
```{r setup}
    knitr::opts_chunk$set(include=FALSE, echo = TRUE, warning = FALSE, message = FALSE)
    library(conflicted)
    library(dendextend)
    library(ggplot2)
    library(here)
    library(sjPlot)
    library(tidyverse)
    library(webshot)
    load(file= here("dat", "lclust-res.RData"))
```

```{r Loading Custom Functions}
    source(here("Functions", "micro.genre.dat.R"))
    source(here("Functions", "demog.dat.R"))
    source(here("Functions", "genre.reg.mods.R"))
    source(here("Functions", "pred.probs.R"))
    source(here("Functions", "plot.preds.R"))
```

```{r Performing Link Clustering}
    lclust.res <- hclust(ldist.res$dist, method = "ward.D2") #clustering of link by link distance matrix
    lclust.dend <- as.dendrogram(lclust.res) #transforming link clustering into dendrogram
```

```{r Microgenre Regression Models}
    spec <- c("educ", "parcoll", "poly(age.n, 2)", 
              "blk", "his", "asn", "gender")
    labs <- c("Intercept", "Education", "Parents College",
                            "Age", "Age-Squared", 
                            "Race ID: Black", "Race ID: Hispanic",
                            "Race ID: Asian", "Gender ID: Woman")
    
    res.metal <- genre.reg.mods("Metal", k = 3, xvars = spec)
    res.salsa <- genre.reg.mods("Salsa", k = 3, xvars = spec)
    
    tab_model(res.metal$mods, 
              digits = 3, show.ci = FALSE, 
              auto.label = FALSE, transform = NULL,
              pred.labels = labs,
              file = here("Tabs", "reg-tab-metal.html"))
    webshot(here("Tabs", "reg-tab-metal.html"), here("Tabs", "reg-tab-metal.png"))
    
    tab_model(res.salsa$mods, 
              digits = 3, show.ci = FALSE, 
              auto.label = FALSE, transform = NULL,
              pred.labels = labs,
              file = here("Tabs", "reg-tab-salsa.html"))
    webshot(here("Tabs", "reg-tab-salsa.html"), here("Tabs", "reg-tab-salsa.png"))
    
```

```{r Microgenre Regression Plots (Age)}
    age.preds <- pred.probs(x = "age.n", res.metal$mods)
    p <- plot.preds(d = age.preds, 
                    xb = c(3, 7, 11),
                    xlabs = c("Young", "Middle Aged", "Old"),
                    ym = -0.02,
                    yl = .2, 
                    xm = 2, 
                    xl = 12)
    save_plot(filename = here("Plots", "Micro", "micro-by-age-metal.png"), 
              fig = p,
              height = 25, width = 20)
    
    age.preds <- pred.probs(x = "age.n", res.salsa$mods)
    p <- plot.preds(d = age.preds, 
                    xb = c(3, 7, 11),
                    xlabs = c("Young", "Middle Aged", "Old"),
                    ym = -0.02,
                    yl = .16, 
                    xm = 2, 
                    xl = 12)
    save_plot(filename = here("Plots", "Micro", "micro-by-age-salsa.png"), 
              fig = p,
              height = 25, width = 20)
```


```{r Microgenre Regression Models (Supp.)}
    spec <- c("educ", "parcoll", "poly(age.n, 2)", 
              "blk", "his", "asn", "gender")
    labs <- c("Intercept", "Education", "Parents College",
                            "Age", "Age-Squared", 
                            "Race ID: Black", "Race ID: Hispanic",
                            "Race ID: Asian", "Gender ID: Woman")
    
    res.folk <- genre.reg.mods("Folk", k = 4, xvars = spec)
    res.indie <- genre.reg.mods("Indie_Alt", k = 4, xvars = spec)
    
    tab_model(res.folk$mods, 
              digits = 3, show.ci = FALSE, 
              auto.label = FALSE, transform = NULL,
              pred.labels = labs,
              file = here("Tabs", "reg-tab-folk.html"))
    #webshot(here("Tabs", "reg-tab-folk.html"), here("Tabs", "reg-tab-folk.png"))
    
    tab_model(res.indie$mods, 
              digits = 3, show.ci = FALSE, 
              auto.label = FALSE, transform = NULL,
              pred.labels = labs,
              file = here("Tabs", "reg-tab-indie.html"))
    #webshot(here("Tabs", "reg-tab-indie.html"), here("Tabs", "reg-tab-indie.png"))
    
```

```{r Microgenre Regression Plots (Age/Supp.)}
    age.preds <- pred.probs(x = "age.n", res.folk$mods)
    p <- plot.preds(d = age.preds, 
                    xb = c(3, 7, 11),
                    xlabs = c("Young", "Middle Aged", "Old"),
                    ym = -0.02,
                    yl = .1, 
                    xm = 2, 
                    xl = 12)
    save_plot(filename = here("Plots", "Micro", "micro-by-age-folk.png"), 
              fig = p,
              height = 25, width = 20)
    
    age.preds <- pred.probs(x = "age.n", res.indie$mods)
    p <- plot.preds(d = age.preds, 
                    xb = c(3, 7, 11),
                    xlabs = c("Young", "Middle Aged", "Old"),
                    ym = -0.05,
                    yl = .25, 
                    xm = 2, 
                    xl = 12)
    save_plot(filename = here("Plots", "Micro", "micro-by-age-indie.png"), 
              fig = p,
              height = 25, width = 20)
```
